backend/requirements.txt
Flask==3.0.3
Flask-CORS==4.0.1
pymongo==4.7.2
bcrypt==4.1.3
PyJWT==2.8.0
python-dotenv==1.0.1
gunicorn==22.0.0
